Move PrpBaseLoader to CommonLoader
Allow this loader to be used by UNPP
